Create Java Fitch template
