+++ JavaFitch/template_makefile
@@ -7,12 +7,12 @@
 
 task1_source_files=*.java
 
-task1_test_files_local=JavaTestRunner.java UnitTest.java tests_Task1.java BTreePrinter.java
+task1_test_files_local=JavaFitchRunner.java UnitTest.java tests_Task1.java
 
 task1_test_files_black_fitch=$(black_fitch_path)/JavaTestRunner.java $(black_fitch_path)/UnitTest.java \
-$(black_fitch_path)/tests_Task1.java $(black_fitch_path)/BTreePrinter.java
+$(black_fitch_path)/tests_Task1.java
 
-task1_main=JavaTestRunner
+task1_main=JavaFitchRunner
 
 task1:
 	@echo "\nCopying Java Fitch..."
