+++ JavaFitch/tests_Task1.java
@@ -1,0 +1,70 @@
+/*
+MIT License
+
+Copyright (c) 2017 Evert Geldenhuys
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+
+public class tests_Task1 extends UnitTest{
+
+
+	public boolean run() {
+
+		boolean result = true;
+		// result = result & tree_single_node("Case: Single Node");
+
+
+
+		return result;
+	}
+
+
+	public boolean test_template(String title) {
+		// HEADER
+		printHeader(title);
+		boolean result = true;
+		boolean subResult = true;
+		// HEADER END
+
+		// =========================================================
+		// TEST CASE START										VVVV
+		printSubTest("Test Template");
+		subResult = true;
+		// ==========================================================
+
+		subresult = subresult & assertCustom("a", "b");
+
+
+		// =========================================================
+		result = result & subResult;
+		printSubTestFooter(subResult);
+		// TEST CASE END										^^^^
+		// =========================================================
+
+
+		// FOOTER
+		printFooter(title, result);
+		return result;
+		// FOOTER END
+
+	}
+
+
+}
