--- assignment2/README.md
@@ -1,40 +1,0 @@
-# COS 212 Assignment 1
-
-## Usage
-### Task 1
-1. `./configure.sh`
-2. Move `black-fitch-makefile` to your source directory:
-
-```
-.
-├── black-fitch-makefile
-├── CircularList.C
-├── CircularList.h
-├── Deque.C
-├── Deque.h
-├── main.C
-├── makefile
-├── package.sh
-├── Queue.C
-├── Queue.h
-├── Stack.C
-└── Stack.h
-```
-
-3. Test with `make -f black-fitch-makefile task1`
-
-### Task 2
-1. `./configure.sh`
-2. Move `black-fitch-makefile` to your source directory:
-```
-.
-├── black-fitch-makefile
-├── CircularList.java
-├── Deque.java
-├── makefile
-├── Queue.java
-├── Stack.java
-└── Tester.java
-```
-
-3. Test with `make -f black-fitch-makefile task2`
